refactor: streamline loading states and remove unused components in affiliate and discount usage pages
