+++ src/components/instructor/affiliate-payout/shared/index.ts
@@ -1,4 +1,3 @@
 export { SearchBar } from "./SearchBar";
 export { FilterBar } from "./FilterBar";
 export { Pagination } from "../../../common/Pagination";
-export { EmptyState } from "./EmptyState";
