+++ src/components/instructor/discount-usage/shared/index.ts
@@ -1,4 +1,3 @@
 export { SearchBar } from "./SearchBar";
 export { FilterBar } from "./FilterBar";
 export { Pagination } from "../../../common/Pagination";
-export { EmptyState } from "./EmptyState";
